Implement TS for Register
Changed React-form-hook to Formik because could not make reusable
FormField component that would serve as password field aswell. There was
somekind of bug/feature that overrides input fields type property from
password to text
